+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/BerryBushes.java
@@ -23,6 +23,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.event.entity.living.LivingAttackEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.registries.RegistryObject;
 
 @Label(name = "Berry Bushes")
@@ -37,14 +38,21 @@ public class BerryBushes extends Feature {
 
     @Config
     @Label(name = "Data pack", description = """
-		Makes sweet berries not plantable, requiring seeds, and also enables a data pack that makes the following changes:
+		Makes sweet berries not plantable, requiring seeds, and also enables a data pack (If berry_good mod is not present) that makes the following changes:
 		* Makes sweet berry bushes drop seeds
 	""")
     public static Boolean dataPack = true;
 
+    /*@Config
+    @Label(name = "Data pack Berry Good", description = """
+		Same as 'Data pack' but uses Berry Good seeds
+	""")
+    public static Boolean dataPackBerryGood = false;*/
+
     public BerryBushes(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
-        IntegratedPack.addPack(new IntegratedPack(PackType.SERVER_DATA, "berries", Component.literal("IguanaTweaks Reborn Berries"), () -> this.isEnabled() && !DataPacks.disableAllDataPacks && dataPack));
+        IntegratedPack.addPack(new IntegratedPack(PackType.SERVER_DATA, "berries", Component.literal("IguanaTweaks Reborn Berries"), () -> this.isEnabled() && !DataPacks.disableAllDataPacks && dataPack && !ModList.get().isLoaded("berry_good")));
+        //IntegratedPack.addPack(new IntegratedPack(PackType.SERVER_DATA, "berries_berry_good", Component.literal("IguanaTweaks Reborn Berries (Berry Good compat)"), () -> this.isEnabled() && !DataPacks.disableAllDataPacks && dataPackBerryGood && ModList.get().isLoaded("berry_good")));
     }
 
     @SubscribeEvent
@@ -62,7 +70,7 @@ public class BerryBushes extends Feature {
     @SubscribeEvent
     public void onTryToPlant(PlayerInteractEvent.RightClickBlock event) {
         if (!this.isEnabled()
-                || !dataPack)
+                || (!dataPack /*&& !dataPackBerryGood*/))
             return;
 
         BlockState stateClicked = event.getLevel().getBlockState(event.getHitVec().getBlockPos());